Move acceptance test setup to their own file
These tests shouldn't be in the Data Source-specific acceptance test
files. There's not really a great place for this stuff, since we have to
share this `*dockertest.Pool` between the tests, and tests don't really
have a way to pass down setup.

Maybe we can find a better way to deal with this in the future. Until
then, this should make the test files a bit more consistent.
