+++ openwrt/internal/system/system_data_source_acceptance_test.go
@@ -5,8 +5,6 @@ package system_test
 import (
 	"context"
 	"fmt"
-	"log"
-	"os"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-framework/providerserver"
@@ -14,34 +12,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/joneshf/terraform-provider-openwrt/internal/acceptancetest"
 	"github.com/joneshf/terraform-provider-openwrt/openwrt"
-	"github.com/ory/dockertest/v3"
 )
-
-var (
-	dockerPool *dockertest.Pool
-)
-
-func TestMain(m *testing.M) {
-	var (
-		code     int
-		err      error
-		tearDown func()
-	)
-	ctx := context.Background()
-	tearDown, dockerPool, err = acceptancetest.Setup(ctx, m)
-	defer func() {
-		tearDown()
-		os.Exit(code)
-	}()
-	if err != nil {
-		fmt.Printf("Problem setting up tests: %s", err)
-		code = 1
-		return
-	}
-
-	log.Printf("Running tests")
-	code = m.Run()
-}
 
 func TestSystemSystemDataSourceAcceptance(t *testing.T) {
 	t.Parallel()
